requirements.txt
asgiref==3.8.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.0
decorator==5.1.1
defusedxml==0.8.0rc2
Django==5.1.5
django-cors==0.1
django-cors-headers==4.6.0
django-extensions==3.2.3
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
future==1.0.0
geocoder==1.38.1
geographiclib==2.0
geopy==2.4.1
idna==3.10
oauthlib==3.2.2
pillow==11.1.0
pycparser==2.22
PyJWT==2.10.1
python3-openid==3.2.0
ratelim==0.1.6
requests==2.32.3
requests-oauthlib==2.0.0
six==1.17.0
social-auth-core==4.5.4
sqlparse==0.5.3
tzdata==2025.1
urllib3==2.3.0
